Add support for python-mode style folding
As a python-mode user in my day-to-day work, I would like the folding
styles of vim-markdown and python-mode to match, so that there's less
cognitive overhead in switching between my two most frequently-used
contexts.

To this end, I've made a distinct set of folding logic, keyed off of the
g:vim_markdown_folding_style_pythonic variable, so that using this style
is opt-in.