Option to not require .md extensions in markdown links
Gitlab uses markdown in an interesting way. If you want to create a link
to the page "linked-to-page" you might do this:

[Linked To Page Title](linked-to-page)

And not,

[Linked To Page Title](linked-to-page.md)

This change is about having the "ge" command work on wiki links
constructed in this way, and provides consistency between how gitlab
works and vim-markdown works.

The default behaviour is not affected.