Add GetHeaderList() in ftplugin/markdown.vim
GetHeaderList() returns a list of headers in the current buffer. This
functionality was part of Toc(), but has been moved to a separate
function for easier reusability.

The returned list contains the text of each header, stripped of
hashes, and the level of each header. Toc() uses the level to
correctly indent headers in the location list window.

As headers are stripped in GetHeaderList() and not in Toc(), some unit
tests in test/folding-toc.vader had to be adapted.