Add `:InsertToc` and `:InsertNToc` commands
The `:InsertToc` command inserts a table of contents at the current
line in the markdown buffer. The `:InsertNToc` command behaves like
`:InsertToc,` but the format of h2 headers in the table of contents is
a numbered list, rather than a bulleted list.

An optional argument can be used to specify how many levels of headers
to display in the table of content, e.g., to display up to and
including h3, use `:InsertToc 3`.

Documentation and unit tests added as well.