Disable code block syntax highlighting when syntax=OFF
When &syntax's value is OFF, vim doesn't load syntax files for the file.
However, this plugin loads code block syntax highlighting rules even if
the value is OFF. This behavior is inconsistent for users to disable
syntax highlighting locally. For example, this causes on a modeline such
as `<!-- vim: set syntax=OFF -->`.